Finished logic for token validation on teams endpoints

+++ src/routes/teams.js
@@ -10,14 +10,36 @@ const main = server => {
   server.get({ path: '/teams', version: '1' }, async (req, res, next) => {
     try {
       const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
-      if (isAdmin) {
+
+      if (!isAdmin) res.send(401, 'Only admins of the API are able to use this endpoint.')
+      else {
         const tService = Mongo.getTeamService(req)
         const teams = await tService.getAllTeams()
 
         res.send(200, teams)
-      } else {
-        res.send(401, 'Only admins of the API are able to use this endpoint.')
       }
+
+      return next()
+    } catch (err) {
+      console.error(err)
+      res.send(err)
+      return next(err)
+    }
+  })
+
+  // Get a list of all the teams created by a specific client.
+  server.get({ path: '/teams/:clientId', version: '1' }, async (req, res, next) => {
+    try {
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
+
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const teams = await tService.getClientTeams()
+
+        res.send(200, teams)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -30,15 +52,15 @@ const main = server => {
   server.post({ path: '/teams/:clientId', version: '1' }, async (req, res, next) => {
     try {
       const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
-      console.log(validToken)
-      // if (isAdmin) {
-      const tService = Mongo.getTeamService(req)
-      const newTeam = await tService.createTeam(req.body)
 
-      res.send(200, newTeam)
-      // } else {
-      //   res.send(401, 'Only admins of the API are able to use this endpoint.')
-      // }
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const newTeam = await tService.createTeam(req.body)
+
+        res.send(200, newTeam)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -50,10 +72,16 @@ const main = server => {
   // Get a specific team.
   server.get({ path: '/teams/:teamId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const team = await tService.getTeam(req.params.teamId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, team)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const team = await tService.getTeam(req.params.teamId)
+
+        res.send(200, team)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -65,10 +93,16 @@ const main = server => {
   // Update an existing team.
   server.put({ path: '/teams/:teamId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const newTeam = await tService.updateTeam(req.params.teamId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, newTeam)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const newTeam = await tService.updateTeam(req.params.teamId, req.body)
+
+        res.send(200, newTeam)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -80,10 +114,16 @@ const main = server => {
   // Delete an existing team.
   server.del({ path: '/teams/:teamId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const oldTeam = await tService.deleteTeam(req.params.teamId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, oldTeam)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const oldTeam = await tService.deleteTeam(req.params.teamId)
+
+        res.send(200, oldTeam)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -95,10 +135,16 @@ const main = server => {
   // Get the list of members currently in the team.
   server.get({ path: '/teams/:teamId/members', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const members = await tService.getMembers(req.params.teamId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, members)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const members = await tService.getMembers(req.params.teamId)
+
+        res.send(200, members)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -110,10 +156,16 @@ const main = server => {
   // Create/add a new member to the team.
   server.post({ path: '/teams/:teamId/members', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const members = await tService.addMember(req.params.teamId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, members)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const members = await tService.addMember(req.params.teamId, req.body)
+
+        res.send(200, members)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -125,10 +177,16 @@ const main = server => {
   // Get the information of a specific team member.
   server.get({ path: '/teams/:teamId/members/:memberId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const member = await tService.getMember(req.params.teamId, req.params.memberId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, member)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const member = await tService.getMember(req.params.teamId, req.params.memberId)
+
+        res.send(200, member)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -140,10 +198,16 @@ const main = server => {
   // Update the information of an existing team member.
   server.put({ path: '/teams/:teamId/members/:memberId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const updatedMember = await tService.updateMember(req.params.teamId, req.params.memberId, req.body)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, updatedMember)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const updatedMember = await tService.updateMember(req.params.teamId, req.params.memberId, req.body)
+
+        res.send(200, updatedMember)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
@@ -155,10 +219,16 @@ const main = server => {
   // Delete an existing team member from the team.
   server.del({ path: '/teams/:teamId/members/:memberId', version: '1' }, async (req, res, next) => {
     try {
-      const tService = Mongo.getTeamService(req)
-      const oldMember = await tService.deleteMember(req.params.teamId, req.params.memberId)
+      const validToken = await validateToken(req.params.clientId, req.headers['request-token'])
 
-      res.send(200, oldMember)
+      if (!validToken) res.send(401, 'The token you provided was either invalid or does not match the provided clientId.')
+      else {
+        const tService = Mongo.getTeamService(req)
+        const oldMember = await tService.deleteMember(req.params.teamId, req.params.memberId)
+
+        res.send(200, oldMember)
+      }
+
       return next()
     } catch (err) {
       console.error(err)
